# spring-openapi-example/src/main/resources/springopenapiexample.yml
openapi: 3.0.1
info:
  version: 1.0.0
  title: Spring openApi example
  description: Spring openApi example
  contact:
    email: techcodediy@gmail.com
tags:
  - name: blog
    description: Operations about blog
    externalDocs:
     description: Find out more about our site
     url: http://swagger.io
paths:
  /blog/{blogtitle}:
    get:
      tags:
        - blog
      summary: Get the blog by title
      operationId: getBlogByTitle
      parameters:
        - name: blogtitle
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Blog'
        404:
          description: Blog not found
          content: { }
  /blog:
    post:
      tags:
      - blog
      summary: "Add a new blog"
      requestBody:
        description: Optional description in *Markdown*
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Blog'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blog'
        404:
          description: Blog not found
          content: { }
    put:
      tags:
        - blog
      summary: "update an existing blog"
      operationId: updateBlog
      requestBody:
        description: Optional description in *Markdown*
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Blog'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blog'
        404:
          description: Blog not found
          content: { }
  /blog/{blogid}:
    delete:
      tags:
        - blog
      summary: Get the blog by title
      operationId: deleteBlog
      parameters:
        - name: blogid
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        404:
          description: Blog not found
          content: { }

components:
  schemas:
    Blog:
      type: object
      properties:
        id:
          type: integer
          format: int64
        blogtitle:
          type: string
        content:
          type: string
        createdate:
          type: string
          format: date
          example: 22-06-2021